+++ crates/matrix-sdk-ui/src/timeline/tests/mod.rs
@@ -29,14 +29,18 @@ use indexmap::IndexMap;
 use matrix_sdk::{
     BoxFuture,
     config::RequestConfig,
-    crypto::{DecryptionSettings, OlmMachine, RoomEventDecryptionResult, TrustRequirement},
     deserialized_responses::{EncryptionInfo, TimelineEvent},
     paginators::{PaginableRoom, PaginatorError, thread::PaginableThread},
     room::{EventWithContextResponse, Messages, MessagesOptions, PushContext, Relations},
     send_queue::RoomSendQueueUpdate,
 };
 use matrix_sdk_base::{
-    RoomInfo, RoomState, crypto::types::events::CryptoContextInfo, latest_event::LatestEvent,
+    RoomInfo, RoomState,
+    crypto::{
+        DecryptionSettings, OlmMachine, RoomEventDecryptionResult, TrustRequirement,
+        types::events::CryptoContextInfo,
+    },
+    latest_event::LatestEvent,
 };
 use matrix_sdk_test::{ALICE, DEFAULT_TEST_ROOM_ID, event_factory::EventFactory};
 use ruma::{